+++ src/context/index.js
@@ -1,11 +1,43 @@
-import unJeune1solution from "./1jeune1solution"
-import mesAides from "./mes-aides.org"
+import Home from "@/views/home.vue"
+import BaseLayout from "@/components/base-layout.vue"
 
-let variation
-if (process.env.VUE_APP_VARIATION === "1jeune1solution") {
-  variation = unJeune1solution
-} else {
-  variation = mesAides
+const routes = [
+  {
+    path: "/accessibilite",
+    name: "accessibilite",
+    component: () =>
+      import(
+        /* webpackChunkName: "accessibilite" */ "@/views/accessibilite.vue"
+      ),
+  },
+  {
+    path: "/contact",
+    name: "contact",
+    component: () =>
+      import(/* webpackChunkName: "contact" */ "@/views/contact.vue"),
+  },
+  {
+    path: "/cgu",
+    name: "cgu",
+    component: () => import(/* webpackChunkName: "cgu" */ "@/views/cgu.vue"),
+  },
+  {
+    path: "/confidentialite",
+    name: "confidentialite",
+    component: () =>
+      import(/* webpackChunkName: "cgu" */ "@/views/confidentialite.vue"),
+  },
+  {
+    path: "/stats",
+    name: "stats",
+    component: () =>
+      import(/* webpackChunkName: "stats" */ "@/views/stats.vue"),
+  },
+]
+
+export default {
+  Home,
+  BaseLayout,
+  name: "1jeune1solution",
+  routes,
 }
-
-export default variation
