--- src/context/mes-aides.org/views/a-propos.vue
@@ -1,118 +1,0 @@
-<template>
-  <article class="text container">
-    <h1>À propos de Mes Aides</h1>
-
-    <h2 id="presentation">Mes Aides, c’est quoi ?</h2>
-
-    <p
-      >Simulez vos droits à plus d’une trentaine de prestations sociales en
-      quelques minutes. Mes Aides compare votre situation personnelle aux
-      critères des principales aides nationales, comme le RSA, la CMU-C ou les
-      allocations familiales, ainsi que de certaines aides locales.</p
-    >
-
-    <p
-      >Grâce aux informations que vous nous transmettez, nous estimons à quelles
-      aides vous avez droit ainsi que leurs montants et conditions
-      d'attribution. Vous pouvez ensuite engager les démarches auprès des
-      organismes ciblés.</p
-    >
-
-    <h2 id="pourquoi">Pourquoi Mes Aides ?</h2>
-
-    <p
-      >Nous luttons contre le non-recours aux prestations sociales. Il s’agit de
-      toutes les situations dans lesquelles une personne est en droit d’obtenir
-      une aide mais ne la reçoit pas. De nombreuses causes sont invoquées comme
-      le manque d’informations ou la complexité des démarches. Avec Mes Aides,
-      nous proposons un résumé clair et synthétique.</p
-    >
-
-    <p
-      >Mes Aides a été lancé en novembre 2014. Depuis janvier 2017, chaque jour,
-      plus de 10 000 usagers évaluent leurs droits à vingt-quatre prestations
-      sociales. Plus de la moitié reçoivent une estimation en moins de sept
-      minutes.</p
-    >
-
-    <small
-      >Explorez les
-      <a
-        href="http://stats.data.gouv.fr/index.php?module=CoreHome&action=index&idSite=9&period=day&date=yesterday#?module=Dashboard&action=embeddedIndex&idSite=9&period=day&date=yesterday&idDashboard=1"
-        >statistiques d'usage</a
-      >
-      et comprenez comment nous évaluons la
-      <a
-        href="https://beta.gouv.fr/2016/11/04/mes-aides-metriques-pirates-tunnel-conversion-public"
-        >performance</a
-      >.</small
-    >
-
-    <h2 id="fonctionnement">Comment fonctionne Mes Aides ?</h2>
-
-    <p
-      >Mes Aides se base sur <a href="https://www.openfisca.fr">OpenFisca</a> :
-      un moteur de calcul libre et ouvert utilisé par des chercheurs en économie
-      et d’autres services publics. À partir des informations que vous entrez
-      dans le simulateur, nous sommes capables d’évaluer votre situation sans
-      avoir à interroger indépendamment chacun des organismes. Nous n’avons pas
-      accès à vos dossiers ni ne transmettons vos informations : à vous de faire
-      les démarches après la simulation.</p
-    >
-
-    <small
-      >Mes Aides ne conserve aucune information nominative : en savoir plus sur
-      <router-link to="/cgu"
-        >l'utilisation de vos données personnelles</router-link
-      >.</small
-    >
-
-    <h2 id="mediateurs"
-      >Puis-je utiliser Mes Aides dans un contexte professionnel ?</h2
-    >
-
-    <p
-      >Notre simulateur est destiné à la fois au grand public et aux médiateurs
-      sociaux.</p
-    >
-
-    <p
-      >En tant que professionnel ou professionnelle de l’action sociale, vous
-      pouvez utiliser Mes Aides pour vous concentrer sur le plus important : la
-      personne en face de vous.</p
-    >
-
-    <small
-      >Approfondissez notre vision de l'<a
-        href="https://hal.archives-ouvertes.fr/hal-01415747/document"
-        >adoption du numérique dans le service social</a
-      >.</small
-    >
-
-    <h2 id="startup-etat">Qui sommes-nous ?</h2>
-
-    <p>Un collectif citoyen.</p>
-
-    <p
-      >Jusqu'en janvier 2020, Mes Aides était une Startup d’État de l’<a
-        href="https://beta.gouv.fr"
-        >Incubateur de services numériques</a
-      >. Un service public numérique produit par une
-      <a href="http://beta.gouv.fr/2016/11/28/equipes-autonomes"
-        >petite équipe autonome</a
-      >. Les pratiques des startups du numérique étaient mises au service de
-      l’action publique.</p
-    >
-
-    <small
-      >Pour en savoir plus, découvrez le programme
-      <a href="https://beta.gouv.fr">beta.gouv.fr</a>, nos
-      <a href="https://github.com/betagouv/mes-aides-ui/wiki/Partenaires"
-        >partenaires</a
-      >, et accédez à tout
-      <a href="https://github.com/betagouv"
-        >le code de l'Incubateur de services numériques</a
-      >.</small
-    >
-  </article>
-</template>
